change lab material
change lab material to coursera lab5,using the explicit free list to
complete.

+++ Chapter9/course-materials/lab5/memlib.c
@@ -24,14 +24,14 @@ static char *mem_max_addr;   /* largest legal heap address */
  */
 void mem_init(void)
 {
-    /* allocate the storage we will use to model the available VM */
-    if ((mem_start_brk = (char *)malloc(MAX_HEAP)) == NULL) {
-	fprintf(stderr, "mem_init_vm: malloc error\n");
-	exit(1);
-    }
+  /* allocate the storage we will use to model the available VM */
+  if ((mem_start_brk = (char *)malloc(MAX_HEAP)) == NULL) {
+    fprintf(stderr, "mem_init_vm: malloc error\n");
+    exit(1);
+  }
 
-    mem_max_addr = mem_start_brk + MAX_HEAP;  /* max legal heap address */
-    mem_brk = mem_start_brk;                  /* heap is empty initially */
+  mem_max_addr = mem_start_brk + MAX_HEAP;  /* max legal heap address */
+  mem_brk = mem_start_brk;                  /* heap is empty initially */
 }
 
 /* 
@@ -39,7 +39,7 @@ void mem_init(void)
  */
 void mem_deinit(void)
 {
-    free(mem_start_brk);
+  free(mem_start_brk);
 }
 
 /*
@@ -47,7 +47,7 @@ void mem_deinit(void)
  */
 void mem_reset_brk()
 {
-    mem_brk = mem_start_brk;
+  mem_brk = mem_start_brk;
 }
 
 /* 
@@ -55,17 +55,17 @@ void mem_reset_brk()
  *    by incr bytes and returns the start address of the new area. In
  *    this model, the heap cannot be shrunk.
  */
-void *mem_sbrk(int incr) 
+void *mem_sbrk(size_t incr) 
 {
-    char *old_brk = mem_brk;
+  char *old_brk = mem_brk;
 
-    if ( (incr < 0) || ((mem_brk + incr) > mem_max_addr)) {
-	errno = ENOMEM;
-	fprintf(stderr, "ERROR: mem_sbrk failed. Ran out of memory...\n");
-	return (void *)-1;
-    }
-    mem_brk += incr;
-    return (void *)old_brk;
+  if ( (incr < 0) || ((mem_brk + incr) > mem_max_addr)) {
+    errno = ENOMEM;
+    fprintf(stderr, "ERROR: mem_sbrk failed. Ran out of memory...\n");
+    return (void *)-1;
+  }
+  mem_brk += incr;
+  return (void *)old_brk;
 }
 
 /*
@@ -73,7 +73,7 @@ void *mem_sbrk(int incr)
  */
 void *mem_heap_lo()
 {
-    return (void *)mem_start_brk;
+  return (void *)mem_start_brk;
 }
 
 /* 
@@ -81,7 +81,7 @@ void *mem_heap_lo()
  */
 void *mem_heap_hi()
 {
-    return (void *)(mem_brk - 1);
+  return (void *)(mem_brk - 1);
 }
 
 /*
@@ -89,7 +89,7 @@ void *mem_heap_hi()
  */
 size_t mem_heapsize() 
 {
-    return (size_t)(mem_brk - mem_start_brk);
+  return (size_t)(mem_brk - mem_start_brk);
 }
 
 /*
@@ -97,5 +97,5 @@ size_t mem_heapsize()
  */
 size_t mem_pagesize()
 {
-    return (size_t)getpagesize();
+  return (size_t)getpagesize();
 }
